Changes to shape objects to explicitly set opaque property when changing colors
When a shapes color or transparency is being changed the proper transparent or opaque properties need to be changed in case it is going from a transparent to opaque state (or vice-versa). Also, in the roundrect __init__ the chunk needs to be drawn even if the initial fill in None; otherwise there seems to be nothing to change to color of later on.